Change logic asyncStorage, add list empty, search and list stocks

+++ src/pages/Home/Home.tsx
@@ -7,6 +7,7 @@ import { DatePickerComponent } from "../../components/DatePicker/DatePicker";
 import { ButtonSubmit } from "../../components/ButtonSubmit/ButtonSubmit";
 import { Input } from "../../components/Input/Input";
 import * as S from "./Home.styles";
+import { Keyboard, TouchableWithoutFeedback } from "react-native";
 
 export function Home() {
   const [idStock, setIdStock] = useState("");
@@ -19,6 +20,10 @@ export function Home() {
 
   const { getItem, setItem } = useAsyncStorage("@mycapitaltest:stocks");
 
+  const handleBlur = () => {
+    Keyboard.dismiss();
+  };
+
   async function handleSubmitStock() {
     try {
       const id = uuid.v4();
@@ -26,23 +31,50 @@ export function Home() {
       const date = selectedDate.toISOString().slice(0, 10);
       const formattedDate = date.split("-").reverse().join("/");
 
-      console.log(formattedDate);
-
       const newData = {
-        id,
-        idStock,
-        nameStock,
-        priceStock,
-        formattedDate
+        formattedDate,
+        items: [
+          {
+            id,
+            idStock,
+            nameStock,
+            priceStock,
+          },
+        ],
       };
 
       const response = await getItem();
       const previousData = response ? JSON.parse(response) : [];
 
-      const data = [...previousData, newData];
+      const existingFormattedData = previousData.filter(
+        (data) => data.formattedDate === formattedDate
+      );
 
-      await setItem(JSON.stringify(data));
-      console.log(JSON.stringify(data));
+      if (existingFormattedData && existingFormattedData.length > 0) {
+        const existingData = existingFormattedData[0];
+
+        const updatedData = {
+          ...existingData,
+          items: [
+            ...existingData.items,
+            {
+              id,
+              idStock,
+              nameStock,
+              priceStock,
+            },
+          ],
+        };
+
+        const newData = previousData.map((data) =>
+          data.formattedDate === formattedDate ? updatedData : data
+        );
+
+        await setItem(JSON.stringify(newData));
+      } else {
+        const data = [...previousData, newData];
+        await setItem(JSON.stringify(data));
+      }
 
       toast.show({
         title: "Ação cadastrada com sucesso!",
@@ -50,7 +82,7 @@ export function Home() {
         bgColor: "green.500",
       });
     } catch (error) {
-      console.log(error);
+      console.log("error", error);
 
       toast.show({
         title: "Algo deu errado, tente novamente mais tarde.",
@@ -63,6 +95,7 @@ export function Home() {
     setNameStock("");
     setPriceStock("");
     setSelectedDate(undefined);
+    handleBlur();
   }
 
   function handleDateSelected(date: Date) {
@@ -74,41 +107,42 @@ export function Home() {
   }, [idStock, nameStock, priceStock, selectedDate]);
 
   return (
-    <S.Container>
-      <S.ContainerTitle>
-        <S.Title>Cadastro de ação</S.Title>
-      </S.ContainerTitle>
+    <TouchableWithoutFeedback onPress={handleBlur}>
+      <S.Container>
+        <S.ContainerTitle>
+          <S.Title>Cadastro de ação</S.Title>
+        </S.ContainerTitle>
 
-      <Input
-        label="Nome da ação"
-        placeholder="ex: Petrobras"
-        onChangeText={setNameStock}
-        value={nameStock}
-        fullWidth
-      />
-      <S.ContainerInput>
         <Input
-          label="Código da ação"
-          maxLength={5}
-          autoCapitalize="characters"
-          placeholder="ex: PETR4"
-          onChangeText={setIdStock}
-          value={idStock}
+          label="Nome da ação"
+          placeholder="ex: Petrobras"
+          onChangeText={setNameStock}
+          value={nameStock}
+          fullWidth
         />
-        <Input
-          isPriceInput={true}
-          label="Valor da ação"
-          placeholder="ex: R$25,96"
-          keyboardType="numeric"
-          onChangeText={setPriceStock}
-          value={priceStock}
-          
-        />
-      </S.ContainerInput>
+        <S.ContainerInput>
+          <Input
+            label="Código da ação"
+            maxLength={5}
+            autoCapitalize="characters"
+            placeholder="ex: PETR4"
+            onChangeText={setIdStock}
+            value={idStock}
+          />
+          <Input
+            isPriceInput={true}
+            label="Valor da ação"
+            placeholder="ex: R$25,96"
+            keyboardType="numeric"
+            onChangeText={setPriceStock}
+            value={priceStock}
+          />
+        </S.ContainerInput>
 
-      <DatePickerComponent onDateSelected={handleDateSelected} />
+        <DatePickerComponent onDateSelected={handleDateSelected} />
 
-      <ButtonSubmit onPress={handleSubmitStock} disabled={!fieldsFilled} />
-    </S.Container>
+        <ButtonSubmit onPress={handleSubmitStock} disabled={!fieldsFilled} />
+      </S.Container>
+    </TouchableWithoutFeedback>
   );
 }
